move to minibatch.py
